Update nargs for input parameters.
with nargs='+' option in add_argument, argparse.ArgumentParser() will pass a parameter as a list. In this case if a parameter is set as "filename.csv", it will be converted to ["filename.csv"], if we pass a list: ["filename.csv"], the list format is preserved.

>>>import argparse
>>> parser = argparse.ArgumentParser()
>>>parser.add_argument(
...       '--train-files',
...       nargs='+',
...       help='Training file local or GCS',
...       default='gs://cloud-samples-data/ml-engine/census/data/adult.data.csv')
>>> parser.parse_args('a b --train-files test.csv')
Namespace(train_files=['test.csv'])
>>> parser.parse_args('a b --train-files [test.csv]')
Namespace(train_files=['[test.csv]'])