This change improve the input_pipeline performance.
This is achieved by calling batching the raw input first and then
calling the map function for each batch. This follows the considerations
described [here](https://www.tensorflow.org/guide/performance/datasets#map_and_batch).

In a related project this change boosted throughput on a CSV dataset by
a factor of 2.

Finally, now the pipeline support skipping the header for every CSV
file. Before it was failed if the CSV dataset consisted of multiple
files each with a header.