This change adds support for the Tensorflow Model Analysis.
This is part of a series of improvements tracked in
GoogleCloudPlatform/cloudml-samples#366 .

This is done by adding 2 new input_receivers_functions (prepended with
tfma_ in the code) and an additional export in task.py.

The template now depends on tfma consequently tf==1.12.0. Therefore,
these dependencies have been updated in setup.py.